refactoring - added new controller for CarType; added log4j2 dependencies;
