+++ src/main/java/pl/edu/wszib/controller/CarController.java
@@ -1,25 +1,39 @@
 package pl.edu.wszib.controller;
 
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import pl.edu.wszib.domain.entity.Car;
-import pl.edu.wszib.domain.entity.CarType;
 import pl.edu.wszib.domain.repository.CarRepository;
-import pl.edu.wszib.domain.repository.CarTypeRepository;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/cars")
 public class CarController {
+    private static final Logger logger = LogManager.getLogger(CarController.class);
 
     @Autowired
     CarRepository carRepository;
 
-    @Autowired
-    CarTypeRepository carTypeRepository;
+    @GetMapping("{/}")
+    public List<Car> getAllCars() {
+        logger.info("retrieving all cars from database");
+        return carRepository.findAll();
+    }
 
     @GetMapping("/{id}")
     public Car getCarById(@PathVariable Integer id) {
+        logger.info("retrieving car with id: " + id);
         return carRepository.getOne(id);
+    }
+
+    @GetMapping("/{registrationnumber}")
+    public Car getCarByRegistrationNumber(@PathVariable String registrationnumber) {
+        logger.info("retrieving car with registration number: " + registrationnumber);
+        return carRepository.findCarByRegistrationNumber(registrationnumber);
     }
 
     @GetMapping("/{cartype}")
@@ -27,34 +41,18 @@ public class CarController {
         return carRepository.findByCarType(cartype);
     }
 
-    @GetMapping("/{brand}/{model}")
-    public CarType getCarByModel(@PathVariable String brand, @PathVariable String model) {
-        return carTypeRepository.findByModel(brand, model);
-    }
 
-    @GetMapping("/{enginecapacity}")
-    public CarType getCarByEngineCapacity(@PathVariable String enginecapacity) {
-        return carTypeRepository.findByEngineCapacity(enginecapacity);
-    }
-
-
-    @GetMapping("/{enginetype}")
-    public CarType getCarByEngineType(@PathVariable String enginetype) {
-        return carTypeRepository.findByEngineType(enginetype);
-    }
-
-    @PostMapping("/{cartype}")
-    public void addTypeOfCar(@PathVariable CarType cartype) {
-        carTypeRepository.save(cartype);
-
-        System.out.println("Car type saved!");
-    }
 
     @PostMapping("/{car}")
     public void addCar(@PathVariable Car car) {
         carRepository.save(car);
 
         System.out.println("car added");
+    }
+
+    @PutMapping("/{id}")
+    public void updateCar(@PathVariable Integer id) {
+        carRepository.getOne(id);
     }
